Move langchain test fakes to apps/utils/tests
Relocate test-support fakes out of the production import path and move
dict_to_json_schema to apps/utils/json. Clears the dubious openai_assistant
lazy-import WARN. Refs #3560.
