+++ apps/pipelines/tests/test_nodes_with_tools.py
@@ -30,7 +30,7 @@ from apps.utils.factories.experiment import (
 )
 from apps.utils.factories.pipelines import PipelineFactory
 from apps.utils.factories.service_provider_factories import LlmProviderFactory, LlmProviderModelFactory
-from apps.utils.langchain import (
+from apps.utils.tests.langchain import (
     build_fake_llm_service,
 )
